requirements.txt
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
asyncio-throttle==1.0.2
backoff==2.2.1 